requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
bcrypt
python-dotenv